requirements.txt
Flask==2.3.0
pyodbc==4.0.34
python-dotenv==1.0.0